Configuration classes
The ConfigLoader takes in the documented config paths and loads in the
values for use by the system.

DropboxSession translates those config values into a properly
initialized dropbox state.

Station translates config values into a ready-to-use object that can be
used to ultimately et up a new countertop for each station.

Issue #1